+++ examples/md5/main.go
@@ -32,7 +32,7 @@ func ProcessMD5(
 		data := make(chan uint32)
 
 		go memory.ReadBurstUInt32(readReqFlit, readRespFlit, inputData,
-			protocol.SmiMemReadOptDefault, uint16(num32s), data)
+			protocol.SmiMemReadOptDefault, uint32(num32s), data)
 
 		for i := numBlocks; i != 0; i-- {
 			for j := 0; j != 16; j += 1 {
